+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.2.0 - 2017-11-06
+
+### Added
+
+* If the second argument of `d.data()` is an array of keys, returns an object with all {key: value}
+* If the second argument of `d.data()` is undefined, returns an object with all values
+
 ## 2.1.0 - 2017-10-20
 
 ### Added